+++ src/data/info_in_position.lua
@@ -1,0 +1,7 @@
+-- 儲存各地點的資訊
+-- 格式為x,y,資訊三個一組
+local INFO_IN_POSITION = {
+    -1793, -2991, "測試",
+}
+
+return INFO_IN_POSITION